TRIPLEX-619 Исправления замечаний по карточке после тестирования
