+++ src/components/Card/styles/Action.module.less
@@ -43,7 +43,7 @@
         outline: none;
 
         &.focusVisible {
-            border: 1px solid var(--triplex-next-Card-BorderColor_Focus);
+            box-shadow: var(--triplex-next-Card-Shadow_Hover), inset var(--triplex-next-Card-Shadow_Focus);
         }
     }
 }
