+++ src/components/Card/styles/Card.module.less
@@ -30,27 +30,11 @@
 
     .cardContent {
         &.paddingMD {
-            padding: 23px;
-
-            &:not(:first-child) {
-                padding-top: 24px;
-            }
-
-            &:not(:last-child) {
-                padding-bottom: 24px;
-            }
+            padding: 24px;
         }
 
         &.paddingSM {
-            padding: 15px;
-
-            &:not(:first-child) {
-                padding-top: 16px;
-            }
-
-            &:not(:last-child) {
-                padding-bottom: 16px;
-            }
+            padding: 16px;
         }
 
         .cardContentHeader {
