+++ stories/Cards/Cards.less
@@ -1,0 +1,12 @@
+.card-static-general-playground-preview,
+.card-playground-preview {
+    padding: 16px;
+    border-radius: 8px;
+    width: 232px;
+}
+
+.card-static-general-playground-preview {
+    .theme-light & {
+        background-color: #d0d7dd;
+    }
+}
